Time: 0 ms (100.00%) | Memory: 19.9 MB (12.13%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -20,8 +20,8 @@ public:
         q.push({root, 0});
         while (!q.empty()) {
             int size = q.size();
-            int first, last;
-            int mmin = q.front().second; // to make id starting from 0
+            long long first, last;
+            long long mmin = q.front().second; // to make id starting from 0
             for (int i = 0; i < size; i++) {
                 long long currIndx = q.front().second - mmin;
                 TreeNode* node = q.front().first;
